+++ test/parallel.js
@@ -7,7 +7,7 @@ import { TEST_VERSION, OLD_TEST_VERSION } from './helpers/versions.js'
 test('Forward exit code and output on late failure | runParallel', async (t) => {
   const { exitCode, stdout, stderr } = await runCli(
     '--parallel --continue',
-    `${TEST_VERSION} ${OLD_TEST_VERSION}`,
+    `${TEST_VERSION},${OLD_TEST_VERSION}`,
     'node -p Buffer.from("")',
   )
 
@@ -27,7 +27,7 @@ Node ${OLD_TEST_VERSION} failed with exit code 1`),
 test('No --continue | runParallel', async (t) => {
   const { exitCode, stdout, stderr } = await runCli(
     '--parallel',
-    `${OLD_TEST_VERSION} ${TEST_VERSION}`,
+    `${OLD_TEST_VERSION},${TEST_VERSION}`,
     'node -p Buffer.from("")',
   )
 
@@ -44,7 +44,7 @@ Node ${OLD_TEST_VERSION} failed with exit code 1`,
 test('--continue | runParallel', async (t) => {
   const { exitCode, stdout, stderr } = await runCli(
     '--parallel --continue',
-    `${OLD_TEST_VERSION} ${TEST_VERSION}`,
+    `${OLD_TEST_VERSION},${TEST_VERSION}`,
     'node -p Buffer.from("")',
   )