+++ test/parse.js
@@ -1,7 +1,8 @@
 import test from 'ava'
 import { each } from 'test-each'
 
-import { runCli } from './helpers/run.js'
+import { runCli, runSerial, runParallel } from './helpers/run.js'
+import { TEST_VERSION } from './helpers/versions.js'
 
 each(
   [
@@ -14,6 +15,17 @@ each(
       const { exitCode } = await runCli('', versionRange, command)
 
       t.is(exitCode, 1)
+    })
+  },
+)
+
+each(
+  [runCli, runSerial, runParallel],
+  ['--fetch', '--no-fetch', '--fetch=true', '--fetch true'],
+  ({ title }, run, opts) => {
+    test(`Parse nve CLI flags | ${title}`, async (t) => {
+      const { exitCode } = await run(opts, TEST_VERSION, 'node --version')
+      t.is(exitCode, 0)
     })
   },
 )