+++ test/serial.js
@@ -6,7 +6,7 @@ import { TEST_VERSION, OLD_TEST_VERSION } from './helpers/versions.js'
 test(`Forward exit code and output on late failure | runSerial`, async (t) => {
   const { exitCode, stdout, stderr } = await runCli(
     '',
-    `${TEST_VERSION} ${OLD_TEST_VERSION}`,
+    `${TEST_VERSION},${OLD_TEST_VERSION}`,
     'node -p Buffer.from("")',
   )
 
@@ -27,7 +27,7 @@ test(`Forward exit code and output on late failure | runSerial`, async (t) => {
 test(`--continue | runSerial`, async (t) => {
   const { exitCode, stdout, stderr } = await runCli(
     '--continue',
-    `${OLD_TEST_VERSION} ${TEST_VERSION}`,
+    `${OLD_TEST_VERSION},${TEST_VERSION}`,
     'node -p Buffer.from("")',
   )